refactor(sources): applies norminette
- logical break in functions
- separate functions into more files
- remove unused functions
- change variable names
