+++ sources/executor/heredoc.c
@@ -6,108 +6,81 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/25 08:12:39 by peda-cos          #+#    #+#             */
-/*   Updated: 2025/04/28 21:35:06 by jlacerda         ###   ########.fr       */
+/*   Updated: 2025/05/02 20:02:58 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "executor.h"
 
-static char	*append_to_buffer(char *buffer, char *line)
-{
-	char	*temp;
-	char	*result;
-
-	temp = buffer;
-	buffer = ft_strjoin(buffer, line);
-	free(temp);
-	if (!buffer)
-		return (NULL);
-	temp = buffer;
-	result = ft_strjoin(buffer, "\n");
-	free(temp);
-	return (result);
-}
-
-static int	is_quoted_delimiter(char *delim)
-{
-	int	len;
-
-	len = ft_strlen(delim);
-	return (len >= 2 && delim[0] == '\'' && delim[len - 1] == '\'');
-}
-
-static char	*process_heredoc_line(char *line, char **env, int last_exit,
-		int expand_vars)
-{
-	char			*expanded;
-	int				has_error;
-	t_expansion_ctx	expansion_cxt;
-
-	has_error = FALSE;
-	if (!expand_vars)
-		return (ft_strdup(line));
-	expansion_cxt.envs = env;
-	expansion_cxt.last_exit = &last_exit;
-	expansion_cxt.has_error_flag_control = &has_error;
-	expanded = expand_variable(line, &expansion_cxt);
-	return (expanded);
-}
-
-static char	*read_heredoc_content(char *delim, char **env, int last_exit)
+static char	*process_heredoc_content(char *stripped_delim,
+	int expand_vars, char **envs, int last_exit)
 {
 	char	*line;
 	char	*buffer;
 	char	*processed_line;
-	int		expand_vars;
-	char	*stripped_delim;
 
 	buffer = ft_strdup("");
 	if (!buffer)
 		return (NULL);
-	expand_vars = !is_quoted_delimiter(delim);
-	if (!expand_vars && delim[0] == '\'')
-	{
-		stripped_delim = ft_substr(delim, 1, ft_strlen(delim) - 2);
-		if (!stripped_delim)
-		{
-			free(buffer);
-			return (NULL);
-		}
-	}
-	else
-		stripped_delim = ft_strdup(delim);
-	if (!stripped_delim)
-	{
-		free(buffer);
-		return (NULL);
-	}
-	while (1)
+	while (TRUE)
 	{
 		line = readline("> ");
 		if (!line || !ft_strcmp(line, stripped_delim))
-		{
-			free(line);
-			free(stripped_delim);
-			break ;
-		}
-		processed_line = process_heredoc_line(line, env, last_exit,
-				expand_vars);
+			return (free(line), buffer);
+		processed_line = process_heredoc_line(line,
+				envs, last_exit, expand_vars);
 		free(line);
 		if (!processed_line)
-		{
-			free(buffer);
-			free(stripped_delim);
-			return (NULL);
-		}
+			return (free(buffer), NULL);
 		buffer = append_to_buffer(buffer, processed_line);
 		free(processed_line);
 		if (!buffer)
-		{
-			free(stripped_delim);
 			return (NULL);
+	}
+}
+
+static char	*read_heredoc_content_to_buffer(char *delim)
+{
+	char	*tmp;
+	char	*line;
+	char	*buffer;
+
+	buffer = ft_strdup("");
+	while (TRUE)
+	{
+		line = readline("> ");
+		if (!line || !ft_strcmp(line, delim))
+		{
+			free(line);
+			break ;
 		}
+		tmp = buffer;
+		buffer = ft_strjoin(buffer, line);
+		free(tmp);
+		tmp = buffer;
+		buffer = ft_strjoin(buffer, "\n");
+		free(tmp);
+		free(line);
 	}
 	return (buffer);
+}
+
+static char	*read_heredoc_content(char *delim, char **env, int last_exit)
+{
+	char	*content;
+	int		expand_vars;
+	char	*stripped_delim;
+
+	expand_vars = !is_quoted_delimiter(delim);
+	stripped_delim = get_stripped_delim(expand_vars, delim);
+	if (!stripped_delim)
+		return (NULL);
+	content = process_heredoc_content(stripped_delim,
+			expand_vars, env, last_exit);
+	free(stripped_delim);
+	if (!content)
+		return (NULL);
+	return (content);
 }
 
 int	handle_heredoc(char *delim, char **env, int last_exit)
@@ -131,37 +104,11 @@ int	handle_heredoc(char *delim, char **env, int last_exit)
 	return (pipefd[0]);
 }
 
-char	*read_heredoc_content_to_buffer(char *delim)
-{
-	char	*line;
-	char	*buffer;
-	char	*tmp;
-
-	buffer = ft_strdup("");
-	while (1)
-	{
-		line = readline("> ");
-		if (!line || !ft_strcmp(line, delim))
-		{
-			free(line);
-			break ;
-		}
-		tmp = buffer;
-		buffer = ft_strjoin(buffer, line);
-		free(tmp);
-		tmp = buffer;
-		buffer = ft_strjoin(buffer, "\n");
-		free(tmp);
-		free(line);
-	}
-	return (buffer);
-}
-
 void	preprocess_heredocs(t_command *cmd_list)
 {
 	t_command	*cmd;
-	int			pipefd[2];
 	char		*buffer;
+	int			pipefd[2];
 
 	cmd = cmd_list;
 	while (cmd)
@@ -169,10 +116,7 @@ void	preprocess_heredocs(t_command *cmd_list)
 		if (cmd->heredoc_delim)
 		{
 			if (pipe(pipefd) < 0)
-			{
-				perror("pipe");
-				return ;
-			}
+				return (perror("pipe"));
 			buffer = read_heredoc_content_to_buffer(cmd->heredoc_delim);
 			if (buffer)
 			{
