+++ sources/tokenizer/token.c
@@ -6,23 +6,11 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/22 13:18:28 by jlacerda          #+#    #+#             */
-/*   Updated: 2025/04/28 21:37:11 by jlacerda         ###   ########.fr       */
+/*   Updated: 2025/05/02 20:44:52 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "tokenizer.h"
-
-t_token	*get_last_token(t_token **tokens)
-{
-	t_token	*tmp;
-
-	if (!*tokens)
-		return (NULL);
-	tmp = *tokens;
-	while (tmp->next)
-		tmp = tmp->next;
-	return (tmp);
-}
 
 t_token_content	*new_token_content(char *value,
 	int in_quotes, int in_single_quotes)
