+++ sources/utils.c
@@ -6,13 +6,13 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 19:05:27 by peda-cos          #+#    #+#             */
-/*   Updated: 2025/04/28 21:19:31 by jlacerda         ###   ########.fr       */
+/*   Updated: 2025/05/02 18:58:43 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static void	set_last_arg_without_pipe_executed(
+void	set_last_arg_without_pipe_executed(
 	t_token *tokens, t_command *cmd, char ***envs)
 {
 	t_token	*current_token;
@@ -96,7 +96,7 @@ void	execute_parsed_commands(t_command *cmd, char ***env, int *last_exit,
 	free_commands(cmd);
 }
 
-static void	process_invalid_command(t_command *cmd,
+void	process_invalid_command(t_command *cmd,
 	int *last_exit, t_token *tokens)
 {
 	if (cmd && cmd->args && *cmd->args[0] == '\0')
@@ -107,27 +107,5 @@ static void	process_invalid_command(t_command *cmd,
 	else
 		*last_exit = 1;
 	free_commands(cmd);
-	free_tokens(tokens);
-}
-
-void	process_input(char *input, char ***env, int *last_exit)
-{
-	t_token		*tokens;
-	t_command	*cmd;
-
-	if (!input || !*input)
-		return ;
-	add_history(input);
-	save_command_to_history(input);
-	g_signal_received = 0;
-	tokens = tokenize_input(input);
-	if (process_tokens(&tokens, last_exit))
-		return ;
-	cmd = parse_tokens(tokens, *env, *last_exit);
-	if (cmd == NULL || cmd->args == NULL || *cmd->args[0] == '\0')
-		return (process_invalid_command(cmd, last_exit, tokens));
-	preprocess_heredocs(cmd);
-	set_last_arg_without_pipe_executed(tokens, cmd, env);
-	execute_parsed_commands(cmd, env, last_exit, tokens);
 	free_tokens(tokens);
 }
